Mega refactoring with substantial induced pain.
This is a megachange (for the project): we're getting back Giovanni's
idea to have separate static/templates directory. It was a good idea.

Probably the most important changes are:

1. Every controller is a blueprint (thanks to Giovanni, again:))

This will prove a very flexible approach as soon as new controllers will
be added to the main app.

2. A app-wide configuration file that will ease the pain: it uses common
defaults so everything can be run without specifying anything among the
env vars.

All the previous env vars still apply and they do override the default
parameters.
