+++ .travis.yml
@@ -22,7 +22,7 @@ before_script:
 install: "pip install -r requirements.txt"
 
 # Run tests.
-script: "python emotly_tests.py"
+script: "python run_tests.py"
 
 # Deploy.
 deploy:
