+++ emotly/controllers/user_controller.py
@@ -1,11 +1,16 @@
+"""
+Emotly
+
+DEED
+"""
 import os
 import bcrypt
 from flask import Blueprint, request, render_template, flash
-from Emotly.models import User
+from emotly import app
+from emotly.models import User
 
-
-user_controller = Blueprint('user_controller',
-                            __name__)
+# User Controller
+user_controller = Blueprint('user_controller', __name__)
 
 
 @user_controller.route("/signup", methods=["GET", "POST"])
@@ -27,9 +32,8 @@ def register_user(req):
     req_pwd = req.form['inputPassword'].encode('utf-8')
     req_email = req.form['inputEmail']
 
-    rnds = int(os.environ['EMOTLY_APP_SEC_ROUNDS']) if\
-        'EMOTLY_APP_SEC_ROUNDS' in os.environ else 12
-    salt = bcrypt.gensalt(rnds)
+    salt = bcrypt.gensalt(app.config['GENSALT_ROUNDS'])
+
     hash_pwd = bcrypt.hashpw(req_pwd, salt)
     user = User(nickname=req_nickname, password=hash_pwd, salt=salt,
                 email=req_email)
