+++ emotly/static/assets/js/ie10-viewport-bug-workaround.js
