+++ emotly/static/emotly.css
