+++ emotly/static/pub/fonts/glyphicons-halflings-regular.eot
