+++ emotly/static/pub/fonts/glyphicons-halflings-regular.ttf
