+++ run.py
@@ -1,0 +1,8 @@
+"""
+Emotly
+
+DEED
+"""
+
+from emotly import app
+app.run()
